+++ src/pages/common/Header/Menu/Profile/Profile.tsx
@@ -3,7 +3,7 @@ import Foco from 'react-foco'
 import type { UserStore } from 'src/stores/User/user.store'
 import { inject, observer } from 'mobx-react'
 import Flex from 'src/components/Flex'
-import { Avatar } from 'src/components/Avatar'
+import { Avatar } from 'src/components/Avatar/Avatar'
 import { ProfileModal } from 'src/components/ProfileModal/ProfileModal'
 import MenuMobileLink from 'src/pages/common/Header/Menu/MenuMobile/MenuMobileLink'
 import ProfileButtons from './ProfileButtons'
